# faiss: use autovectorization for inner product