--- a/tests/test_contrib_with_scipy.py
+++ b/tests/test_contrib_with_scipy.py
@@ -56,13 +56,13 @@ class TestClustering(unittest.TestCase):
         D, I = clustering.sparse_assign_to_dense(xsparse, centroids)
 
         np.testing.assert_array_equal(Iref.ravel(), I)
-        np.testing.assert_array_almost_equal(Dref.ravel(), D, decimal=4)
+        np.testing.assert_array_almost_equal(Dref.ravel(), D, decimal=3)
 
         D, I = clustering.sparse_assign_to_dense_blocks(
             xsparse, centroids, qbs=123, bbs=33, nt=4)
 
         np.testing.assert_array_equal(Iref.ravel(), I)
-        np.testing.assert_array_almost_equal(Dref.ravel(), D, decimal=4)
+        np.testing.assert_array_almost_equal(Dref.ravel(), D, decimal=3)
 
     def test_sparse_kmeans(self):
         """ demo on how to cluster sparse data into dense clusters """