@zcfh 2025-04-25
> > May I ask which benchmark you are running?
> 
> Sorry for not publishing benchmarks, it's not terribly fancy. It's something like this: https://godbolt.org/z/Y1oj7xsjo Quick Bench doesn't seem to work at the moment, or I'd share that.

Are there any requirements for the compiler version?
I used clang-17 but could not reproduce the optimization effect.