@zcfh 2025-04-29
> > May I ask which benchmark you are running?
> 
> Sorry for not publishing benchmarks, it's not terribly fancy. It's something like this: https://godbolt.org/z/Y1oj7xsjo Quick Bench doesn't seem to work at the moment, or I'd share that.

What compiler version and compile options are you using?