@zcfh 2025-04-30
>  Is there something you are trying to do?
`fvec_inner_product` is a hot function in my program. I am looking for optimization points for it, and I saw this commit mentioning that using automatic vectorization can bring better performance. I only improved by 40% on the local test benchmark, so I want to know if my compiler or compilation options are wrong.

> To the best of my knowledge, this codegen: https://godbolt.org/z/9e3TMozqc is pretty good
Compiler and the previous code do roughly the same thing but compiler unrolls - which is important in high IPC code like this.

It seems that no special options are used, so I will try to analyze the reason myself.