@DenisYaroshevskiy 2025-04-30
The specific improvement will vary by the machine.
Are you using march=native or some similar option?